Added the ability to create multiple levels
Added a level before the existing one.
Made most things prefabs to help with level making.
Each level has a level info object. Might not be the most efficient way of storing metadata but it works.

And finally... replaced the playable character with obamium. (also removed obama cube but it's a prefab)